requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
